requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Deep Learning & AI
torch>=1.9.0
transformers>=4.11.0
spacy>=3.1.0
tensorrt>=8.0.0

# Speech Processing
deepgram-sdk>=2.0.0
sounddevice>=0.4.0
pyaudio>=0.2.11

# Cloud Services
openai>=0.27.0
azure-cognitiveservices-speech>=1.19.0
azure-storage-blob>=12.9.0

# Vehicle Integration
python-can>=3.3.4
mbux-sdk>=1.0.0  # Custom MBUX SDK

# Stream Processing
apache-flink>=1.14.0
pyflink>=1.14.0

# Security
noise-protocol>=0.1.0
cryptography>=3.4.0

# Testing
pytest>=6.2.5
pytest-asyncio>=0.15.1
pytest-cov>=2.12.0

# Development
black>=21.7b0
flake8>=3.9.0
mypy>=0.910 